Fix #9281: qlik sense col. lineage, code reduced
